requirements.txt
absl-py==2.1.0
addict==2.4.0
anykeystore==0.2
apex==0.9.10.dev0
cachetools==5.3.2
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cryptacular==1.6.2
cycler==0.12.1
defusedxml==0.7.1
einops==0.7.0
facexlib==0.3.0
filelock==3.13.1
filterpy==1.4.5
fonttools==4.47.2
fsspec==2023.12.2
ftfy==6.1.3
future==0.18.3
google-auth==2.27.0
google-auth-oauthlib==1.2.0
greenlet==3.0.3
grpcio==1.60.0
huggingface-hub==0.20.3
hupper==1.12.1
idna==3.6
imageio==2.33.1
imgaug==0.4.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
Jinja2==3.1.3
kiwisolver==1.4.5
lazy_loader==0.3
llvmlite==0.41.1
lmdb==1.4.1
lpips==0.1.4
Markdown==3.5.2
MarkupSafe==2.1.4
matplotlib==3.8.2
mpmath==1.3.0
networkx==3.2.1
numba==0.58.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.3.101
nvidia-nvtx-cu12==12.1.105
oauthlib==3.2.2
openai-clip==1.0.1
opencv-python==4.9.0.80
opencv-python-headless==4.9.0.80
packaging==23.2
pandas==2.2.0
PasteDeploy==3.1.0
pbkdf2==1.3
pillow==10.2.0
plaster==1.1.2
plaster-pastedeploy==1.0.1
platformdirs==4.1.0
protobuf==4.23.4
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyiqa==0.1.10
pyparsing==3.1.1
pyramid==2.0.2
pyramid-mailer==0.15.1
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
repoze.sendmail==4.4.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
safetensors==0.4.2
scikit-image==0.22.0
scipy==1.12.0
shapely==2.0.2
six==1.16.0
SQLAlchemy==2.0.25
sympy==1.12
tensorboard==2.15.1
tensorboard-data-server==0.7.2
tifffile==2023.12.9
timm==0.9.12
tomli==2.0.1
torch==2.1.2
torchgeometry==0.1.2
torchvision==0.16.2
tqdm==4.66.1
transaction==4.0
translationstring==1.4
triton==2.1.0
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.1.0
velruse==1.1.1
venusian==3.1.0
wcwidth==0.2.13
WebOb==1.8.7
Werkzeug==3.0.1
WTForms==3.1.2
wtforms-recaptcha==0.3.2
yapf==0.40.2
zipp==3.17.0
zope.deprecation==5.0
zope.interface==6.1
zope.sqlalchemy==3.1
